db관련 수정

+++ server/src/db/migrations/018_add_job_role_code_to_employees.js
@@ -5,18 +5,21 @@
  */
 
 exports.up = async function (knex) {
-  await knex.schema.alterTable('employees', (table) => {
-    // 공통코드 그룹 JOB_ROLE의 code 값을 저장
-    // 예: BACKEND_DEV, FULLSTACK_DEV, DBA, UIUX_DESIGNER
-    table.string('job_role_code', 50).nullable();
-
-    table.index(['job_role_code'], 'employees_job_role_code_idx');
-  });
+	const hasColumn = await knex.schema.hasColumn('employees', 'job_role_code');
+	if (!hasColumn) {
+		await knex.schema.alterTable('employees', (table) => {
+			table.string('job_role_code', 50).nullable();
+			table.index(['job_role_code'], 'employees_job_role_code_idx');
+		});
+	}
 };
 
 exports.down = async function (knex) {
-  await knex.schema.alterTable('employees', (table) => {
-    table.dropIndex(['job_role_code'], 'employees_job_role_code_idx');
-    table.dropColumn('job_role_code');
-  });
+	const hasColumn = await knex.schema.hasColumn('employees', 'job_role_code');
+	if (hasColumn) {
+		await knex.schema.alterTable('employees', (table) => {
+			table.dropIndex(['job_role_code'], 'employees_job_role_code_idx');
+			table.dropColumn('job_role_code');
+		});
+	}
 };
